pyproject.toml
[project]
name = "mailassist"
version = "0.1.0"
description = "LLM-powered mail processing job"
authors = [{name = "MailAssist Team"}]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "pypdf>=4.0.0",
    "PyYAML>=6.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
